backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.4.5
sqlalchemy==2.0.25
pydantic[email]>=2.6.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
stripe==7.12.0
httpx==0.26.0
python-dotenv==1.0.0
